스킬 기능 추가 준비4
스킬 기능을 위한 뼈대 만드는 중
스킬 사용 부분은 완성
스킬을 얻는 부분 해야 함

+++ ClickerGame/Assets/Scripts/Data/Data.Contents.cs
@@ -30,7 +30,7 @@ namespace Data
         public string skillName;
         public float skillValue;
         public float skillIncreaseValue;
-        public int skillCoolTime;
+        public float skillCoolTime;
     }
     #endregion
